Fix some minor errors
This commit fix some small issues with citron which were generating a
log of logs on console, like missing images, unneeded log messages and
an issue with thorn returning a 401 when the used send invalid
credentials. The right thing to do is return 201, since the
communication did succeeded. This was also fixed on Thorn.